api/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-auth
python-dotenv